[src] Share several NSLayout* implementations between AppKit and UIKit.
