+++ src/Foundation/NSLayoutConstraint.cs
@@ -109,6 +109,27 @@ namespace UIKit
 		{
 			return NSLayoutConstraint.Create (view1, attribute1, relation, null, NSLayoutAttribute.NoAttribute, 1.0f, 0f);
 		}
+
+		// This solves the duplicate selector export problem while not breaking the API.
+		public static NSLayoutConstraint Create (NSObject view1, NSLayoutAttribute attribute1, NSLayoutRelation relation, 
+				NSObject view2, NSLayoutAttribute attribute2, nfloat multiplier, nfloat constant)
+		{
+			return Create ((INativeObject) view1, attribute1, relation, view2, attribute2, multiplier, constant);
+		}
+
+#if !MONOMAC || XAMCORE_4_0
+		[iOS (10, 0)]
+		public NSLayoutAnchor<AnchorType> FirstAnchor<AnchorType> () where AnchorType : NSObject
+		{
+			return Runtime.GetNSObject<NSLayoutAnchor<AnchorType>> (_FirstAnchor ());
+		}
+
+		[iOS (10, 0)]
+		public NSLayoutAnchor<AnchorType> SecondAnchor<AnchorType> () where AnchorType : NSObject
+		{
+			return Runtime.GetNSObject<NSLayoutAnchor<AnchorType>> (_SecondAnchor ());
+		}
+#endif // !MONOMAC || XAMCORE_4_0
 	}
 }
