+++ src/uikit.cs
@@ -360,174 +360,6 @@ namespace UIKit {
 
 #if !WATCH
 	[NoWatch]
-	[iOS (9,0)]
-	[BaseType (typeof(NSObject))]
-	[DisableDefaultCtor] // Handle is nil
-	interface NSLayoutAnchor<AnchorType> : NSCopying, NSCoding
-	{
-		[Export ("constraintEqualToAnchor:")]
-		NSLayoutConstraint ConstraintEqualTo (NSLayoutAnchor<AnchorType> anchor);
-	
-		[Export ("constraintGreaterThanOrEqualToAnchor:")]
-		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor);
-	
-		[Export ("constraintLessThanOrEqualToAnchor:")]
-		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor);
-	
-		[Export ("constraintEqualToAnchor:constant:")]
-		NSLayoutConstraint ConstraintEqualTo (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
-	
-		[Export ("constraintGreaterThanOrEqualToAnchor:constant:")]
-		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
-	
-		[Export ("constraintLessThanOrEqualToAnchor:constant:")]
-		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
-	}
-
-	[NoWatch]
-	[iOS (9,0)]
-	[TV (10,0)]
-	[BaseType (typeof(NSLayoutAnchor<NSLayoutXAxisAnchor>))]
-	[DisableDefaultCtor] // Handle is nil
-	interface NSLayoutXAxisAnchor
-	{
-		[iOS (10,0)]
-		[Export ("anchorWithOffsetToAnchor:")]
-		NSLayoutDimension CreateAnchorWithOffset (NSLayoutXAxisAnchor otherAnchor);
-
-		[TV (11,0), iOS (11,0)]
-		[Export ("constraintEqualToSystemSpacingAfterAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintEqualToSystemSpacingAfterAnchor (NSLayoutXAxisAnchor anchor, nfloat multiplier);
-
-		[TV (11,0), iOS (11,0)]
-		[Export ("constraintGreaterThanOrEqualToSystemSpacingAfterAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintGreaterThanOrEqualToSystemSpacingAfterAnchor (NSLayoutXAxisAnchor anchor, nfloat multiplier);
-
-		[TV (11,0), iOS (11,0)]
-		[Export ("constraintLessThanOrEqualToSystemSpacingAfterAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintLessThanOrEqualToSystemSpacingAfterAnchor (NSLayoutXAxisAnchor anchor, nfloat multiplier);
-	}
-
-	[NoWatch]
-	[iOS (9,0)]
-	[TV (10,0)]
-	[BaseType (typeof(NSLayoutAnchor<NSLayoutYAxisAnchor>))]
-	[DisableDefaultCtor] // Handle is nil
-	interface NSLayoutYAxisAnchor
-	{
-		[iOS (10,0)]
-		[Export ("anchorWithOffsetToAnchor:")]
-		NSLayoutDimension CreateAnchorWithOffset (NSLayoutYAxisAnchor otherAnchor);
-
-		[TV (11,0), iOS (11,0)]
-		[Export ("constraintEqualToSystemSpacingBelowAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintEqualToSystemSpacingBelowAnchor (NSLayoutYAxisAnchor anchor, nfloat multiplier);
-
-		[TV (11,0), iOS (11,0)]
-		[Export ("constraintGreaterThanOrEqualToSystemSpacingBelowAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintGreaterThanOrEqualToSystemSpacingBelowAnchor (NSLayoutYAxisAnchor anchor, nfloat multiplier);
-
-		[TV (11,0), iOS (11,0)]
-		[Export ("constraintLessThanOrEqualToSystemSpacingBelowAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintLessThanOrEqualToSystemSpacingBelowAnchor (NSLayoutYAxisAnchor anchor, nfloat multiplier);
-	}
-
-	[NoWatch]
-	[iOS (9,0)]
-	[BaseType (typeof(NSLayoutAnchor<NSLayoutDimension>))]
-	[DisableDefaultCtor] // Handle is nil
-	interface NSLayoutDimension
-	{
-		[Export ("constraintEqualToConstant:")]
-		NSLayoutConstraint ConstraintEqualTo (nfloat constant);
-	
-		[Export ("constraintGreaterThanOrEqualToConstant:")]
-		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (nfloat constant);
-	
-		[Export ("constraintLessThanOrEqualToConstant:")]
-		NSLayoutConstraint ConstraintLessThanOrEqualTo (nfloat constant);
-	
-		[Export ("constraintEqualToAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintEqualTo (NSLayoutDimension anchor, nfloat multiplier);
-	
-		[Export ("constraintGreaterThanOrEqualToAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier);
-	
-		[Export ("constraintLessThanOrEqualToAnchor:multiplier:")]
-		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier);
-	
-		[Export ("constraintEqualToAnchor:multiplier:constant:")]
-		NSLayoutConstraint ConstraintEqualTo (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
-	
-		[Export ("constraintGreaterThanOrEqualToAnchor:multiplier:constant:")]
-		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
-	
-		[Export ("constraintLessThanOrEqualToAnchor:multiplier:constant:")]
-		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
-	}
-
-	[NoWatch]
-	[BaseType (typeof (NSObject))]
-	interface NSLayoutConstraint {
-		[Static]
-		[Export ("constraintsWithVisualFormat:options:metrics:views:")]
-		NSLayoutConstraint [] FromVisualFormat (string format, NSLayoutFormatOptions formatOptions, [NullAllowed] NSDictionary metrics, NSDictionary views);
-
-		[Static]
-		[Export ("constraintWithItem:attribute:relatedBy:toItem:attribute:multiplier:constant:")]
-		NSLayoutConstraint Create (INativeObject view1, NSLayoutAttribute attribute1, NSLayoutRelation relation, [NullAllowed] INativeObject view2, NSLayoutAttribute attribute2, nfloat multiplier, nfloat constant);
-
-		[Export ("priority")]
-		float Priority { get; set;  } // Returns a float, not nfloat.
-
-		[Export ("shouldBeArchived")]
-		bool ShouldBeArchived { get; set;  }
-
-		[NullAllowed, Export ("firstItem", ArgumentSemantic.Assign)]
-		NSObject FirstItem { get;  }
-
-		[Export ("firstAttribute")]
-		NSLayoutAttribute FirstAttribute { get;  }
-
-		[Export ("relation")]
-		NSLayoutRelation Relation { get;  }
-
-		[Export ("secondItem", ArgumentSemantic.Assign)]
-		NSObject SecondItem { get;  }
-
-		[Export ("secondAttribute")]
-		NSLayoutAttribute SecondAttribute { get;  }
-
-		[Export ("multiplier")]
-		nfloat Multiplier { get;  }
-
-		[Export ("constant")]
-		nfloat Constant { get; set;  }
-
-		[iOS (8,0)]
-		[Export ("active")]
-		bool Active { [Bind ("isActive")] get; set; }
-
-		[iOS (8,0)]
-		[Static, Export ("activateConstraints:")]
-		void ActivateConstraints (NSLayoutConstraint [] constraints);
-
-		[iOS (8,0)]
-		[Static, Export ("deactivateConstraints:")]
-		void DeactivateConstraints (NSLayoutConstraint [] constraints);
-
-		[iOS (10,0), TV (10,0)]
-		[Export ("firstAnchor", ArgumentSemantic.Copy)]
-		[Internal]
-		IntPtr _FirstAnchor<AnchorType> ();
-	
-		[iOS (10,0), TV (10,0)]
-		[Export ("secondAnchor", ArgumentSemantic.Copy)]
-		[Internal]
-		IntPtr _SecondAnchor<AnchorType> ();
-	}
-
-	[NoWatch]
 	[iOS (7,0)] // Yup, it is declared as appearing in 7.0, even if it shipped with 8.0
 	[Category, BaseType(typeof(NSLayoutConstraint))]
 	interface NSIdentifier {
