+++ src/xkit.cs
@@ -1797,4 +1797,297 @@ namespace UIKit {
 		string RepresentedElementKind { get; }
 	}
 
+	[NoWatch]
+	[iOS (9,0)]
+	[Mac (10,11)]
+	[MacCatalyst (13,0)]
+	[BaseType (typeof(NSObject))]
+	[DisableDefaultCtor] // Handle is nil
+	interface NSLayoutAnchor<AnchorType> : NSCopying, NSCoding
+	{
+		[Export ("constraintEqualToAnchor:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintEqualToAnchor (NSLayoutAnchor<AnchorType> anchor);
+#else
+		NSLayoutConstraint ConstraintEqualTo (NSLayoutAnchor<AnchorType> anchor);
+#endif
+
+		[Export ("constraintGreaterThanOrEqualToAnchor:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintGreaterThanOrEqualToAnchor (NSLayoutAnchor<AnchorType> anchor);
+#else
+		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor);
+#endif
+
+		[Export ("constraintLessThanOrEqualToAnchor:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintLessThanOrEqualToAnchor (NSLayoutAnchor<AnchorType> anchor);
+#else
+		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor);
+#endif
+
+		[Export ("constraintEqualToAnchor:constant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintEqualToAnchor (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
+#else
+		NSLayoutConstraint ConstraintEqualTo (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
+#endif
+
+		[Export ("constraintGreaterThanOrEqualToAnchor:constant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintGreaterThanOrEqualToAnchor (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
+#else
+		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
+#endif
+
+		[Export ("constraintLessThanOrEqualToAnchor:constant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintLessThanOrEqualToAnchor (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
+#else
+		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutAnchor<AnchorType> anchor, nfloat constant);
+#endif
+
+		[NoiOS][NoMacCatalyst][NoTV][NoWatch]
+		[Mac (10, 12)]
+		[Export ("name")]
+		string Name { get; }
+
+		[NoiOS][NoMacCatalyst][NoTV][NoWatch]
+		[Mac (10, 12)]
+		[NullAllowed, Export ("item", ArgumentSemantic.Weak)]
+		NSObject Item { get; }
+
+		[NoiOS][NoMacCatalyst][NoTV][NoWatch]
+		[Mac (10, 12)]
+		[Export ("hasAmbiguousLayout")]
+		bool HasAmbiguousLayout { get; }
+
+		[NoiOS][NoMacCatalyst][NoTV][NoWatch]
+		[Mac (10, 12)]
+		[Export ("constraintsAffectingLayout")]
+		NSLayoutConstraint[] ConstraintsAffectingLayout { get; }
+	}
+
+	[NoWatch]
+	[iOS (9,0)]
+	[TV (10,0)]
+	[Mac (10,11)]
+	[MacCatalyst (13,0)]
+	[BaseType (typeof(NSLayoutAnchor<NSLayoutXAxisAnchor>))]
+	[DisableDefaultCtor] // Handle is nil
+	interface NSLayoutXAxisAnchor
+	{
+		[iOS (10,0)]
+		[Mac (10,12)]
+		[Export ("anchorWithOffsetToAnchor:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutDimension GetAnchorWithOffset (NSLayoutXAxisAnchor otherAnchor);
+#else
+		NSLayoutDimension CreateAnchorWithOffset (NSLayoutXAxisAnchor otherAnchor);
+#endif
+
+		[TV (11,0), iOS (11,0)]
+		[Mac (11,0)]
+		[Export ("constraintEqualToSystemSpacingAfterAnchor:multiplier:")]
+		NSLayoutConstraint ConstraintEqualToSystemSpacingAfterAnchor (NSLayoutXAxisAnchor anchor, nfloat multiplier);
+
+		[TV (11,0), iOS (11,0)]
+		[Mac (11,0)]
+		[Export ("constraintGreaterThanOrEqualToSystemSpacingAfterAnchor:multiplier:")]
+		NSLayoutConstraint ConstraintGreaterThanOrEqualToSystemSpacingAfterAnchor (NSLayoutXAxisAnchor anchor, nfloat multiplier);
+
+		[TV (11,0), iOS (11,0)]
+		[Mac (11,0)]
+		[Export ("constraintLessThanOrEqualToSystemSpacingAfterAnchor:multiplier:")]
+		NSLayoutConstraint ConstraintLessThanOrEqualToSystemSpacingAfterAnchor (NSLayoutXAxisAnchor anchor, nfloat multiplier);
+	}
+
+	[NoWatch]
+	[iOS (9,0)]
+	[TV (10,0)]
+	[Mac (10,11)]
+	[MacCatalyst (13,0)]
+	[BaseType (typeof(NSLayoutAnchor<NSLayoutYAxisAnchor>))]
+	[DisableDefaultCtor] // Handle is nil
+	interface NSLayoutYAxisAnchor
+	{
+		[iOS (10,0)]
+		[Mac (10,12)]
+		[Export ("anchorWithOffsetToAnchor:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutDimension GetAnchorWithOffset (NSLayoutYAxisAnchor otherAnchor);
+#else
+		NSLayoutDimension CreateAnchorWithOffset (NSLayoutYAxisAnchor otherAnchor);
+#endif
+
+		[TV (11,0), iOS (11,0)]
+		[Mac (11,0)]
+		[Export ("constraintEqualToSystemSpacingBelowAnchor:multiplier:")]
+		NSLayoutConstraint ConstraintEqualToSystemSpacingBelowAnchor (NSLayoutYAxisAnchor anchor, nfloat multiplier);
+
+		[TV (11,0), iOS (11,0)]
+		[Mac (11,0)]
+		[Export ("constraintGreaterThanOrEqualToSystemSpacingBelowAnchor:multiplier:")]
+		NSLayoutConstraint ConstraintGreaterThanOrEqualToSystemSpacingBelowAnchor (NSLayoutYAxisAnchor anchor, nfloat multiplier);
+
+		[TV (11,0), iOS (11,0)]
+		[Mac (11,0)]
+		[Export ("constraintLessThanOrEqualToSystemSpacingBelowAnchor:multiplier:")]
+		NSLayoutConstraint ConstraintLessThanOrEqualToSystemSpacingBelowAnchor (NSLayoutYAxisAnchor anchor, nfloat multiplier);
+	}
+
+	[NoWatch]
+	[iOS (9,0)]
+	[Mac (10,11)]
+	[BaseType (typeof(NSLayoutAnchor<NSLayoutDimension>))]
+	[DisableDefaultCtor] // Handle is nil
+	interface NSLayoutDimension
+	{
+		[Export ("constraintEqualToConstant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintEqualToConstant (nfloat constant);
+#else
+		NSLayoutConstraint ConstraintEqualTo (nfloat constant);
+#endif
+
+		[Export ("constraintGreaterThanOrEqualToConstant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintGreaterThanOrEqualToConstant (nfloat constant);
+#else
+		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (nfloat constant);
+#endif
+
+		[Export ("constraintLessThanOrEqualToConstant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintLessThanOrEqualToConstant (nfloat constant);
+#else
+		NSLayoutConstraint ConstraintLessThanOrEqualTo (nfloat constant);
+#endif
+
+		[Export ("constraintEqualToAnchor:multiplier:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintEqualToAnchor (NSLayoutDimension anchor, nfloat multiplier);
+#else
+		NSLayoutConstraint ConstraintEqualTo (NSLayoutDimension anchor, nfloat multiplier);
+#endif
+
+		[Export ("constraintGreaterThanOrEqualToAnchor:multiplier:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintGreaterThanOrEqualToAnchor (NSLayoutDimension anchor, nfloat multiplier);
+#else
+		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier);
+#endif
+
+		[Export ("constraintLessThanOrEqualToAnchor:multiplier:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintLessThanOrEqualToAnchor (NSLayoutDimension anchor, nfloat multiplier);
+#else
+		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier);
+#endif
+
+		[Export ("constraintEqualToAnchor:multiplier:constant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintEqualToAnchor (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
+#else
+		NSLayoutConstraint ConstraintEqualTo (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
+#endif
+
+		[Export ("constraintGreaterThanOrEqualToAnchor:multiplier:constant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintGreaterThanOrEqualToAnchor (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
+#else
+		NSLayoutConstraint ConstraintGreaterThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
+#endif
+
+		[Export ("constraintLessThanOrEqualToAnchor:multiplier:constant:")]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutConstraint ConstraintLessThanOrEqualToAnchor (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
+#else
+		NSLayoutConstraint ConstraintLessThanOrEqualTo (NSLayoutDimension anchor, nfloat multiplier, nfloat constant);
+#endif
+	}
+
+	[NoWatch]
+	[MacCatalyst (13,0)]
+	[BaseType (typeof (NSObject))]
+	interface NSLayoutConstraint
+#if MONOMAC
+		: NSAnimatablePropertyContainer
+#endif
+{
+		[Static]
+		[Export ("constraintsWithVisualFormat:options:metrics:views:")]
+		NSLayoutConstraint [] FromVisualFormat (string format, NSLayoutFormatOptions formatOptions, [NullAllowed] NSDictionary metrics, NSDictionary views);
+
+		[Static]
+		[Export ("constraintWithItem:attribute:relatedBy:toItem:attribute:multiplier:constant:")]
+		NSLayoutConstraint Create (INativeObject view1, NSLayoutAttribute attribute1, NSLayoutRelation relation, [NullAllowed] INativeObject view2, NSLayoutAttribute attribute2, nfloat multiplier, nfloat constant);
+
+		[Export ("priority")]
+		float Priority { get; set;  } // Returns a float, not nfloat.
+
+		[Export ("shouldBeArchived")]
+		bool ShouldBeArchived { get; set;  }
+
+		[NullAllowed, Export ("firstItem", ArgumentSemantic.Assign)]
+		NSObject FirstItem { get;  }
+
+		[Export ("firstAttribute")]
+		NSLayoutAttribute FirstAttribute { get;  }
+
+		[Export ("relation")]
+		NSLayoutRelation Relation { get;  }
+
+		[Export ("secondItem", ArgumentSemantic.Assign)]
+		[NullAllowed]
+		NSObject SecondItem { get;  }
+
+		[Export ("secondAttribute")]
+		NSLayoutAttribute SecondAttribute { get;  }
+
+		[Export ("multiplier")]
+		nfloat Multiplier { get;  }
+
+		[Export ("constant")]
+		nfloat Constant { get; set;  }
+
+		[iOS (8,0)]
+		[Mac (10,10)]
+		[Export ("active")]
+		bool Active { [Bind ("isActive")] get; set; }
+
+		[iOS (8,0)]
+		[Mac (10,10)]
+		[Static, Export ("activateConstraints:")]
+		void ActivateConstraints (NSLayoutConstraint [] constraints);
+
+		[iOS (8,0)]
+		[Mac (10,10)]
+		[Static, Export ("deactivateConstraints:")]
+		void DeactivateConstraints (NSLayoutConstraint [] constraints);
+
+		[Mac (10, 12)]
+		[iOS (10,0), TV (10,0)]
+		[Export ("firstAnchor", ArgumentSemantic.Copy)]
+#if MONOMAC && !XAMCORE_4_0
+		NSLayoutAnchor<NSObject> FirstAnchor { get; }
+#else
+		[Internal]
+		IntPtr _FirstAnchor<AnchorType> ();
+#endif
+
+		[Mac (10, 12)]
+		[iOS (10,0), TV (10,0)]
+		[Export ("secondAnchor", ArgumentSemantic.Copy)]
+#if MONOMAC && !XAMCORE_4_0
+		[NullAllowed]
+		NSLayoutAnchor<NSObject> SecondAnchor { get; }
+#else
+		[Internal]
+		IntPtr _SecondAnchor<AnchorType> ();
+#endif
+
+		[NullAllowed, Export ("identifier")]
+		string Identifier { get; set; }
+	}
 }
